requirements.txt
Django==4.1.7
python-dotenv==1.0.0
mysqlclient==2.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
django-cors-headers
geopy
stripe==5.4.0
twilio==8.5.0
xhtml2pdf==0.2.11